requirements.txt
fastapi==0.135.3
uvicorn==0.44.0
sqlalchemy==2.0.49
psycopg2-binary==2.9.11
passlib[bcrypt]==1.7.4
python-jose==3.5.0
python-multipart==0.0.24
pydantic==2.12.5
email-validator==2.3.0
bcrypt==4.2.0
